+++ cmk/gui/graphing/_html_render.py
@@ -370,7 +370,7 @@ def _show_graph_add_to_icon_for_popup(
     )  # Ensures that graph canvas does not cover it
 
 
-def _show_graph_canvas(size: tuple[int, int]) -> None:
+def _show_graph_canvas(size: tuple[float, float]) -> None:
     """Create canvas where actual graph will be rendered"""
     graph_width: float = size[0] * html_size_per_ex
     graph_height: float = size[1] * html_size_per_ex
@@ -607,8 +607,8 @@ def _render_ajax_graph(context: Mapping[str, Any]) -> dict[str, Any]:
 
     if resize_x_var is not None and resize_y_var is not None:
         render_opt_x, render_opt_y = graph_render_config.size
-        size_x = int(max(min_resize_width, float(resize_x_var) / html_size_per_ex + render_opt_x))
-        size_y = int(max(min_resize_height, float(resize_y_var) / html_size_per_ex + render_opt_y))
+        size_x = max(min_resize_width, float(resize_x_var) / html_size_per_ex + render_opt_x)
+        size_y = max(min_resize_height, float(resize_y_var) / html_size_per_ex + render_opt_y)
         user.save_file("graph_size", (size_x, size_y))
         graph_render_config.size = (size_x, size_y)
 
@@ -934,7 +934,7 @@ def _render_time_range_selection(
 
 def make_graph_data_range(
     time_range: tuple[int, int],
-    width_in_ex: int,
+    width_in_ex: float,
 ) -> GraphDataRange:
     return GraphDataRange(
         time_range=time_range,
@@ -944,7 +944,7 @@ def make_graph_data_range(
 
 def estimate_graph_step_for_html(
     time_range: tuple[int, int],
-    width_in_ex: int,
+    width_in_ex: float,
 ) -> int:
     steps_per_ex = html_size_per_ex * 4
     number_of_steps = width_in_ex * steps_per_ex
@@ -1057,10 +1057,10 @@ def host_service_graph_dashlet_cmk(
     graph_display_id: str = "",
 ) -> HTML | None:
     width_var = request.get_float_input_mandatory("width", 0.0)
-    width = int(width_var / html_size_per_ex)
+    width = width_var / html_size_per_ex
 
     height_var = request.get_float_input_mandatory("height", 0.0)
-    height = int(height_var / html_size_per_ex)
+    height = height_var / html_size_per_ex
 
     bounds = _graph_margin_ex(graph_render_config.show_margin)
     height -= _graph_title_height_ex(graph_render_config)
